upgrde to .Net 9 Aspir 9 rc 1

+++ src/PhotoSearch.API/Endpoints/PhotoRetrieval/GetImageEndpoint.cs
@@ -24,11 +24,11 @@ public class GetImageEndpoint(IMongoCollection<Photo> collection): Endpoint<Mode
         using var image = new MagickImage(Path.Combine("../PhotoSearch.Worker/", photo.ExactPath));
 
         if(image.Width>r.MaxWidth)
-            image.Resize(r.MaxWidth, 0);
+            image.Resize((uint)r.MaxWidth, 0);
         if (image.Height>r.MaxHeight)
-            image.Resize(0, r.MaxHeight);
+            image.Resize(0, (uint)r.MaxHeight);
         
-        image.Resize(r.MaxWidth, 0);
+        image.Resize((uint)r.MaxWidth, 0);
         var bytes = image.ToByteArray();
         await using Stream stream = new MemoryStream(bytes);
         await SendStreamAsync(stream, "photo.jpg", bytes.Length, "image/jpeg", cancellation: c);
